refactor: move some utils
