+++ yosoi/integrations/opencode.py
@@ -14,8 +14,8 @@ from pydantic_ai.profiles import ModelProfile
 from pydantic_ai.settings import ModelSettings
 from pydantic_ai.usage import RequestUsage
 
-from yosoi.integrations.messages import flatten_messages
-from yosoi.integrations.usage import build_request_usage
+from yosoi.integrations.utils.messages import flatten_messages
+from yosoi.integrations.utils.usage import build_request_usage
 from yosoi.utils import observability as obs
 
 
@@ -153,7 +153,7 @@ def _usage_from_info(info: dict[str, Any]) -> RequestUsage:
     OpenCode reports usage under ``info.tokens`` as
     ``{input, output, reasoning, cache: {read, write}}``. Returns a zeroed
     ``RequestUsage`` when the server omits ``tokens`` (e.g. an error or an older
-    server build). See :mod:`yosoi.integrations.usage` for why this matters.
+    server build). See :mod:`yosoi.integrations.utils.usage` for why this matters.
     """
     tokens = info.get('tokens')
     if not isinstance(tokens, dict):
